requirements.txt
selenium>=4.9.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pyautogui>=0.9.54
pynput>=1.7.6
Pillow>=10.0.0